refactor(app): separate main and lambda bootstrap points
