+++ src/main.ts
@@ -1,39 +1,21 @@
 import { NestFactory } from '@nestjs/core';
-import { ExpressAdapter } from '@nestjs/platform-express';
-import serverlessExpress from '@vendia/serverless-express';
-import { Callback, Context, Handler } from 'aws-lambda';
-import express from 'express';
 
 import helmet from 'helmet';
 
 import { AppModule } from './app.module';
 
-let cachedServer: Handler;
+const port = process.env.PORT || 4000;
 
 async function bootstrap() {
-  const expressApp = express();
-
-  const app = await NestFactory.create(AppModule,
-    new ExpressAdapter(expressApp),
-);
-  await app.init();
+  const app = await NestFactory.create(AppModule);
 
   app.enableCors({
     origin: (req, callback) => callback(null, true),
   });
-  
   app.use(helmet());
 
-  cachedServer = serverlessExpress({ app: expressApp });
-
-  return cachedServer;
+  await app.listen(port);
 }
-
-export const handler: Handler = async (
-  event: any,
-  context: Context,
-  callback: Callback,
-) => {
-  const server = await bootstrap();
-  return server(event, context, callback);
-};
+bootstrap().then(() => {
+  console.log('App is running on %s port', port);
+});
